pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shamcash"
version = "1.1.2"
description = "Python client for the ShamCash HTTP API (shamcash-api.com)"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "melchersman", email = "melchersman@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
]
keywords = ["shamcash", "payment", "api", "syria", "wallet"]
dependencies = ["aiohttp>=3.8.0"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "aioresponses>=0.7.4",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/Melchersman/shamcash"
Documentation = "https://shamcash-api.com"
Repository = "https://github.com/Melchersman/shamcash"
Issues = "https://github.com/Melchersman/shamcash/issues"

[tool.setuptools]
packages = ["shamcash"]

[tool.setuptools.package-data]
shamcash = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: calls the live ShamCash API (requires SHAMCASH_API_TOKEN)",
]
